Fix score endpoint float validation + tectonic missing libs on Lambda
- app.py: ScoreResponse fields int → float (fixes Pydantic 500 on avg_score)
- Dockerfile.lambda: install graphite2, harfbuzz, fontconfig, freetype, libpng
  (fixes tectonic "libgraphite2.so.3 not found" on Lambda)

+++ app.py
@@ -157,10 +157,10 @@ class ScoreRequest(BaseModel):
 
 
 class ScoreResponse(BaseModel):
-    ats_score: int
-    hiring_manager_score: int
-    tech_recruiter_score: int
-    avg_score: int
+    ats_score: float
+    hiring_manager_score: float
+    tech_recruiter_score: float
+    avg_score: float
     reasoning: str
     matched_resume: str
